hotfix-handle-no-nav-item: handle rendering SubNavigation with a null item
