+++ src/components/Form/Radio/tests/Default.test.js
@@ -1,5 +1,20 @@
 // @flow
-import fixture from '../fixtures/Default.fixture.js';
+import createTestContext from 'react-cosmos-test/enzyme';
 import testFixture from '../../../../../config/testFixture';
+import fixture from '../fixtures/Default.fixture.js';
 
 testFixture(fixture, 'Default');
+
+test('Radio should change when clicked', () => {
+  const { mount, getWrapper } = createTestContext({ fixture });
+
+  mount();
+
+  const wrapper = getWrapper();
+
+  const first = wrapper.find('.option').at(0);
+
+  expect(first.hasClass('selected')).toBe(false);
+
+  first.simulate('click');
+});
