+++ src/components/Navigation/SubNavigation/SubNavigation.tests.js
@@ -1,0 +1,5 @@
+// @flow
+import testFixtures from '../../../../config/testFixture';
+import fixtures from './SubNavigation.fixtures';
+
+testFixtures(fixtures);
