--- src/components/Navigation/SubNavigation/fixtures/Default.fixture.js
@@ -1,39 +1,0 @@
-// @flow
-import SubNavigation from '../';
-
-export default {
-  component: SubNavigation,
-  url: '/feedback',
-  props: {
-    item: {
-      key: 'PARENT',
-      title: 'Parent title',
-      label: 'Parent',
-      route: '/parent',
-      badge: 'FREE',
-      items: [
-        {
-          key: 'MAIN',
-          label: 'Main',
-          route: '/parent',
-        },
-        {
-          key: 'SECONDARY',
-          label: 'Secondary',
-          route: '/parent/secondary',
-        },
-      ],
-    },
-    locations: [
-      {
-        value: 'Location A',
-        label: 'Location A',
-      },
-      {
-        value: 'Location B',
-        label: 'Location B',
-      },
-    ],
-    logoutRoute: '/logout',
-  },
-};
