--- src/components/Navigation/SubNavigation/fixtures/loading.fixture.js
@@ -1,18 +1,0 @@
-// @flow
-import SubNavigation from '../';
-
-export default {
-  component: SubNavigation,
-  url: '/feedback',
-  props: {
-    item: {
-      items: [],
-      key: 'placeholder',
-      title: '',
-      route: '#',
-    },
-    locations: [],
-    loading: true,
-    logoutRoute: '/logout',
-  },
-};
