--- src/components/Navigation/SubNavigation/tests/Default.test.js
@@ -1,5 +1,0 @@
-// @flow
-import fixture from '../fixtures/Default.fixture.js';
-import testFixture from '../../../../../config/testFixture';
-
-testFixture(fixture, 'Default');
