--- src/components/Navigation/SubNavigation/tests/loading.test.js
@@ -1,5 +1,0 @@
-// @flow
-import fixture from '../fixtures/loading.fixture.js';
-import testFixture from '../../../../../config/testFixture';
-
-testFixture(fixture, 'loading');
