--- src/components/Navigation/SubNavigation/tests/noChildren.test.js
@@ -1,5 +1,0 @@
-// @flow
-import fixture from '../fixtures/noChildren.fixture.js';
-import testFixture from '../../../../../config/testFixture';
-
-testFixture(fixture, 'noChildren');
